# storage-service/target/generated-sources/openapi/src/main/java/com/example/physiological/storage/api/StorageApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.physiological.storage.api;

import com.example.physiological.storage.model.CustomerRequest;
import com.example.physiological.storage.model.HospitalRequest;
import com.example.physiological.storage.model.StorageRequest;
import com.example.physiological.storage.model.WardRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-04T22:28:53.467019400+05:30[Asia/Colombo]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Storage", description = "the Storage API")
public interface StorageApi {

    /**
     * POST /storage/customer : Create a new customer
     *
     * @param customerRequest  (required)
     * @return Customer created successfully (status code 200)
     *         or Invalid input (status code 400)
     */
    @Operation(
        operationId = "createCustomer",
        summary = "Create a new customer",
        tags = { "Storage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Customer created successfully"),
            @ApiResponse(responseCode = "400", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/storage/customer",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> createCustomer(
        @Parameter(name = "CustomerRequest", description = "", required = true) @Valid @RequestBody CustomerRequest customerRequest
    );


    /**
     * POST /{customer}/storage/hospital : Create a new hospital
     *
     * @param customer  (required)
     * @param hospitalRequest  (required)
     * @return Hospital created successfully (status code 200)
     *         or Invalid input (status code 400)
     */
    @Operation(
        operationId = "createHospital",
        summary = "Create a new hospital",
        tags = { "Storage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Hospital created successfully"),
            @ApiResponse(responseCode = "400", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/{customer}/storage/hospital",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> createHospital(
        @Parameter(name = "customer", description = "", required = true, in = ParameterIn.PATH) @PathVariable("customer") String customer,
        @Parameter(name = "HospitalRequest", description = "", required = true) @Valid @RequestBody HospitalRequest hospitalRequest
    );


    /**
     * POST /{customer}/storage/ward : Create a new ward
     *
     * @param customer  (required)
     * @param wardRequest  (required)
     * @return Ward created successfully (status code 200)
     *         or Invalid input (status code 400)
     */
    @Operation(
        operationId = "createWard",
        summary = "Create a new ward",
        tags = { "Storage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Ward created successfully"),
            @ApiResponse(responseCode = "400", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/{customer}/storage/ward",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> createWard(
        @Parameter(name = "customer", description = "", required = true, in = ParameterIn.PATH) @PathVariable("customer") String customer,
        @Parameter(name = "WardRequest", description = "", required = true) @Valid @RequestBody WardRequest wardRequest
    );


    /**
     * POST /{customer}/storage/data : Store physiological data
     *
     * @param customer The customer identifier (required)
     * @param storageRequest  (required)
     * @return Data stored successfully (status code 200)
     *         or Invalid input (status code 400)
     */
    @Operation(
        operationId = "storeData",
        summary = "Store physiological data",
        tags = { "Storage" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Data stored successfully"),
            @ApiResponse(responseCode = "400", description = "Invalid input")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/{customer}/storage/data",
        consumes = { "application/json" }
    )
    
    ResponseEntity<Void> storeData(
        @Parameter(name = "customer", description = "The customer identifier", required = true, in = ParameterIn.PATH) @PathVariable("customer") String customer,
        @Parameter(name = "StorageRequest", description = "", required = true) @Valid @RequestBody StorageRequest storageRequest
    );

}